requirements.txt
fastapi
uvicorn
openai
python-dotenv
faiss-cpu
numpy
rank-bm25
pydantic